makaneyyat_last/requirements.txt
arrow==0.10.0
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
diff-match-patch==20121119
Django==1.11.5
django-cors-headers==2.2.0
django-filter==1.1.0
django-haystack==2.8.1
django-hstore==1.4.2
django-import-export==0.5.1
django-modelcluster==3.1
django-taggit==0.22.1
django-taggit-serializer==0.1.7
django-treebeard==4.1.2
djangorestframework==3.8.2
djangorestframework-gis==0.11.2
drf-haystack==1.8.3
et-xmlfile==1.0.1
Flask==0.12.2
geos==0.2.1
html5lib==0.999999999
idna==2.6
itsdangerous==0.24
jdcal==1.3
Jinja2==2.9.6
kazoo==2.2.1
lxml==4.0.0
MarkupSafe==1.0
odfpy==1.3.5
olefile==0.44
openpyxl==2.4.8
Pillow==4.3.0
pkg-resources==0.0.0
proj==0.1.0
psycopg2==2.7.3.1
pysolr==3.6.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.18.4
six==1.11.0
SolrClient==0.2.1
tablib==0.12.1
unicodecsv==0.14.1
Unidecode==0.4.21
urllib3==1.22
wagtail==1.12.2
webencodings==0.5.1
Werkzeug==0.12.2
Willow==0.4
xlrd==1.1.0
xlwt==1.3.0
